# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /result-checker/check:
    post:
      operationId: checkResult
      tags: [student]
      summary: Check a result using an examination number and EPIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResultCheckInput"
      responses:
        "200":
          description: Checked result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultCheckResponse"
...
  /exams:
    get:
      operationId: listExams
      tags: [student]
      summary: List available examinations
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, upcoming, live, completed]
            default: all
      responses:
        "200":
          description: Examinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
  /exams/{examId}:
    get:
      operationId: getExam
      tags: [student]
      summary: Get an examination and its questions
      parameters:
        - $ref: "#/components/parameters/ExamId"
      responses:
        "200":
          description: Examination details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamDetail"
  /exams/{examId}/attempts:
    post:
      operationId: startAttempt
      tags: [student]
      summary: Start or resume an examination attempt
      parameters:
        - $ref: "#/components/parameters/ExamId"
      responses:
        "201":
          description: Examination attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attempt"
  /attempts/{attemptId}:
    get:
      operationId: getAttempt
      tags: [student]
      summary: Get a current attempt
      parameters:
        - $ref: "#/components/parameters/AttemptId"
      responses:
        "200":
          description: Attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attempt"
  /attempts/{attemptId}/answers:
    patch:
      operationId: saveAttemptAnswer
      tags: [student]
      summary: Save one answer in an attempt
      parameters:
        - $ref: "#/components/parameters/AttemptId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "200":
          description: Updated attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attempt"
  /attempts/{attemptId}/submit:
    post:
      operationId: submitAttempt
      tags: [student]
      summary: Submit an examination attempt and calculate its score
      parameters:
        - $ref: "#/components/parameters/AttemptId"
      responses:
        "200":
          description: Examination result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
...
  /admin/school-info:
    patch:
      operationId: updateSchoolInfo
      tags: [admin]
      summary: Update school information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolInfoInput"
      responses:
        "200":
          description: Updated school information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolInfo"
  /admin/epins:
    get:
      operationId: listEpins
      tags: [admin]
      summary: List result checker EPINs
      responses:
        "200":
          description: EPINs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Epin"
    post:
      operationId: generateEpins
      tags: [admin]
      summary: Generate one or more result checker EPINs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EpinGenerateInput"
      responses:
        "201":
          description: Generated EPINs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Epin"
  /admin/epins/{epinId}:
    patch:
      operationId: updateEpin
      tags: [admin]
      summary: Disable or reactivate an EPIN
      parameters:
        - $ref: "#/components/parameters/EpinId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EpinUpdate"
      responses:
        "200":
          description: Updated EPIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Epin"
...
  /admin/subjects:
    get:
      operationId: listSubjects
      tags: [admin]
      summary: List subjects
      responses:
        "200":
          description: Subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [admin]
      summary: Create a subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /admin/subjects/{subjectId}:
    patch:
      operationId: updateSubject
      tags: [admin]
      summary: Update a subject
      parameters:
        - $ref: "#/components/parameters/SubjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [admin]
      summary: Delete a subject
      parameters:
        - $ref: "#/components/parameters/SubjectId"
      responses:
        "204":
          description: Subject deleted
  /admin/questions:
    get:
      operationId: listQuestions
      tags: [admin]
      summary: List questions in the question bank
      parameters:
        - name: subjectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [admin]
      summary: Create a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /admin/questions/{questionId}:
    patch:
      operationId: updateQuestion
      tags: [admin]
      summary: Update a question
      parameters:
        - $ref: "#/components/parameters/QuestionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [admin]
      summary: Delete a question
      parameters:
        - $ref: "#/components/parameters/QuestionId"
      responses:
        "204":
          description: Question deleted
  /admin/exams:
    get:
      operationId: listAdminExams
      tags: [admin]
      summary: List examinations for management
      responses:
        "200":
          description: Examinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: createExam
      tags: [admin]
      summary: Create an examination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamInput"
      responses:
        "201":
          description: Created examination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /admin/exams/{examId}:
    patch:
      operationId: updateExam
      tags: [admin]
      summary: Update an examination
      parameters:
        - $ref: "#/components/parameters/ExamId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamInput"
      responses:
        "200":
          description: Updated examination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    delete:
      operationId: deleteExam
      tags: [admin]
      summary: Delete an examination
      parameters:
        - $ref: "#/components/parameters/ExamId"
      responses:
        "204":
          description: Examination deleted
...
components:
  parameters:
    ExamId:
      name: examId
      in: path
      required: true
      schema:
        type: integer
...
    AttemptId:
      name: attemptId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Exam:
      type: object
      required: [id, title, subjectId, subjectName, dateLabel, durationMinutes, questionCount, status, color]
      properties:
        id: { type: integer }
        title: { type: string }
        subjectId: { type: integer }
        subjectName: { type: string }
        dateLabel: { type: string }
        startAt: { type: ["string", "null"], format: date-time }
        durationMinutes: { type: integer }
        questionCount: { type: integer }
        status:
          type: string
          enum: [upcoming, live, completed, draft]
        color: { type: string }
        averageScore: { type: ["number", "null"] }
...
    ExamDetail:
      type: object
      required: [exam, questions]
      properties:
        exam:
          $ref: "#/components/schemas/Exam"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
...
    Attempt:
      type: object
      required: [id, examId, startedAt, expiresAt, status, answers]
      properties:
        id: { type: integer }
        examId: { type: integer }
        startedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        status:
          type: string
          enum: [in_progress, submitted, expired]
        answers:
          type: object
          additionalProperties:
            type: integer
...
    ResultCheckInput:
      type: object
      required: [examinationNumber, epin]
      properties:
        examinationNumber: { type: string, minLength: 1 }
        epin: { type: string, minLength: 6 }
    ResultCheckResponse:
      type: object
      required: [school, result]
      properties:
        school:
          $ref: "#/components/schemas/SchoolInfo"
        result:
          $ref: "#/components/schemas/Result"